--- src/modules/App/App.actions.test.ts
@@ -1,12 +1,0 @@
-import * as SUT from './App.actions'
-import { AppActions } from './types'
-
-describe('modules/App/App.actions', () => {
-  describe('resetAppData', () => {
-    it('should return correct action type', async () => {
-      const state = { type: AppActions.resetApp }
-      const actions = SUT.resetAppData()
-      expect(actions).toEqual(state)
-    })
-  })
-})
