--- src/modules/App/App.actions.ts
@@ -1,5 +1,0 @@
-import { createAction } from '@reduxjs/toolkit'
-
-import { AppActions } from './types'
-
-export const resetAppData = createAction(AppActions.resetApp)
