+++ src/modules/App/App.middleware.test.ts
@@ -1,27 +1,28 @@
 import { createMiddlewareMock } from '../../../tests/tests.utils'
-import { resetAppData } from './App.actions'
 import * as SUT from './App.middleware'
+import { resetAppData } from './App.reducer'
 
 describe('modules/App/App.middleware', () => {
   describe('appResetFlow', () => {
-    it('should correctly call app reset action', async () => {
+    it('should correctly handle an app reset action', () => {
       const create = createMiddlewareMock(jest)
       // given ... the resetAppData action is fired
       const action = resetAppData()
-      const { next, invoke } = create(SUT.appResetFlow)
+      const { store, next, invoke } = create(SUT.appResetFlow)
 
-      await invoke(action)
+      invoke(action)
       // ... we validate that our actions were triggered
       expect(next).toHaveBeenCalledWith(action)
+      expect(store.dispatch).toHaveBeenCalled()
     })
-    it('should handle app reset action accurately', async () => {
+    it('should handle app reset action accurately', () => {
       const create = createMiddlewareMock(jest)
 
       // given ... the resetAppData action is fired
       const action = resetAppData()
       const { store, invoke } = create(SUT.appResetFlow)
 
-      await invoke(action)
+      invoke(action)
       // ... we validate that our actions were triggered
       expect(store.dispatch).toBeCalled()
     })
