+++ src/modules/App/App.middleware.ts
@@ -1,13 +1,13 @@
 import { Middleware } from 'redux'
 
 import * as AuthActions from './../Auth/Auth.reducer'
-import { resetAppData } from './App.actions'
+import { resetAppData } from './App.reducer'
 
-export const appResetFlow: Middleware = ({ dispatch }) => next => async action => {
+export const appResetFlow: Middleware = ({ dispatch }) => next => action => {
   const result = next(action)
 
   if (resetAppData.match(action)) {
-    dispatch(AuthActions.clearAuthState)
+    dispatch(AuthActions.clearAuth())
   }
 
   return result
