+++ src/modules/App/App.reducer.ts
@@ -1,0 +1,5 @@
+import { createAction } from '@reduxjs/toolkit'
+
+const name = '[App]'
+
+export const resetAppData = createAction(`${name} Clear App Reducers`)
