--- src/modules/App/types.ts
@@ -1,3 +1,0 @@
-export enum AppActions {
-  resetApp = '[App] Clear App Reducers',
-}
